in-line `TestContext` into SemanticSyntaxCheckerVisitor
